move user creation from users to auth controller
