+++ controllers/Auth.Controller.js
@@ -3,14 +3,20 @@ const jsonwebtoken = require('jsonwebtoken');
 const bcrypt = require('bcrypt');
 const ApiResponse = require('../common/ApiResponse');
 const LoginsService = require('../services/Logins.Service');
+const UsersService = require('../services/Users.Service');
 
 // api/v0/auth
 class AuthController {
   constructor() {
     this.router = Router();
     this.loginsService = new LoginsService();
+    this.usersService = new UsersService();
 
     this.router.post('/login', this.login);
+    this.router.post('/register', this.register);
+    // TBD
+    // this.router.post('/users/:userId/password', this.sendEmailResetPwd);
+    // this.router.post('/reset-password', this.resetPassword);
   }
 
   login = async (req, res) => {
@@ -56,7 +62,61 @@ class AuthController {
     return res.status(apiResponse.statusCode).json(apiResponse);
   };
 
-  // password reset?
+  register = async (req, res) => {
+    const apiResponse = new ApiResponse();
+    try {
+      const { username, email, dateOfBirth, password } = req.body;
+
+      // Step find username if already exist
+      const getUserResponse = await this.loginsService.findByUsername(username);
+      if (getUserResponse.result) {
+        apiResponse.conflict('This username already exists');
+        return res.response(apiResponse);
+      }
+
+      // TODO
+      // validate request body here
+
+      // Step generate hashed password
+      const salt = await bcrypt.genSalt();
+      const passwordHash = await bcrypt.hash(password, salt);
+
+      // Step create login account
+      const login = { username, passwordHash };
+      const createLoginResponse = await this.loginsService.create(login);
+      if (createLoginResponse.fields.length) {
+        apiResponse.badRequest(
+          'You have some errors',
+          createLoginResponse.fields
+        );
+        return res.response(apiResponse);
+      }
+      if (!createLoginResponse.result) {
+        apiResponse.unprocessableEntity(
+          'Login account cannot be created, please try again later'
+        );
+        return res.response(apiResponse);
+      }
+
+      // Step create user profile
+      const user = { username, email, dateOfBirth };
+      const createUserResponse = await this.usersService.add(user);
+      if (createUserResponse.fields.length) {
+        apiResponse.badRequest('Check for errors', createUserResponse.fields);
+        return res.response(apiResponse);
+      }
+      if (!createUserResponse.result) {
+        apiResponse.unprocessableEntity(
+          'User cannot be created, please try again later'
+        );
+        return res.response(apiResponse);
+      }
+      apiResponse.created(createUserResponse);
+    } catch (error) {
+      apiResponse.internalServerError(error);
+    }
+    return res.response(apiResponse);
+  };
 }
 
 module.exports = AuthController;
