+++ controllers/Users.Controller.js
@@ -38,52 +38,6 @@ class UsersController {
     return res.status(apiResponse.statusCode).json(apiResponse);
   };
 
-  post = async (req, res) => {
-    const apiResponse = new ApiResponse();
-    try {
-      const { username, email, dateOfBirth, password } = req.body;
-      const getUserResponse = await this.loginsService.findByUsername(username);
-      if (getUserResponse.result) {
-        apiResponse.conflict('This username already exists');
-        return res.status(apiResponse.statusCode).json(apiResponse);
-      }
-
-      const salt = await bcrypt.genSalt();
-      const passwordHash = await bcrypt.hash(password, salt);
-      const login = { username, passwordHash };
-      const createLoginResponse = await this.loginsService.create(login);
-      if (createLoginResponse.fields.length) {
-        apiResponse.badRequest(
-          'You have some errors',
-          createLoginResponse.fields
-        );
-        return res.status(apiResponse.statusCode).json(apiResponse);
-      }
-      if (!createLoginResponse.result) {
-        apiResponse.unprocessableEntity(
-          'Login account cannot be created, please try again later'
-        );
-        return res.status(apiResponse.statusCode).json(apiResponse);
-      }
-      const user = { username, email, dateOfBirth };
-      const createUserResponse = await this.usersService.add(user);
-      if (createUserResponse.fields.length) {
-        apiResponse.badRequest('Check for errors', createUserResponse.fields);
-        return res.status(apiResponse.statusCode).json(apiResponse);
-      }
-      if (!createUserResponse.result) {
-        apiResponse.unprocessableEntity(
-          'User cannot be created, please try again later'
-        );
-        return res.status(apiResponse.statusCode).json(apiResponse);
-      }
-      apiResponse.created(createUserResponse);
-    } catch (error) {
-      apiResponse.internalServerError(error);
-    }
-    return res.status(apiResponse.statusCode).json(apiResponse);
-  };
-
   getById = async (req, res) => {
     const apiResponse = new ApiResponse();
     try {
